// MobileAPP/src/utilis/types/Registration.ts
export interface editProfile {
  height: string;
  physicalStatus: string;
  weight: string;
  bodyType: string;
  interCommunity: string;
  educationInstitution: string;
  Organization: string;
  smokingHabits: string;
  drinkingHabits: any;
  dietHabits: any;
  fatherOccupation: any;
  motherOccupation: any;
  education: string;
  employedIn: string;
  occupation: string;
  currency: any;
  income: any;
  familyStatus: any;
  familyType: any;
  familyValue: any;
  ancestralOrigin: string;
  aboutMyself: string;
  fatherName: string;
  motherName: string;
  maleCount: string;
  femaleCount: string;
  maleMarriedCount: string;
  femaleMarriedCount: string;
  location: any;
  city: any;
  state: any;
  address: string;

  educationInstitutionError: any;
  OrganizationError: string;
  smokingHabitsError: string;
  drinkingHabitsError: string;
  dietHabitsError: string;
  fatherOccupationError: string;
  motherOccupationError: string;
  educationError: string;
  employedInError: string;
  occupationError: string;
  incomeError: string;
  familyStatusError: string;
  familyTypeError: string;
  familyValueError: string;
  ancestralOriginError: string;
  aboutMyselfError: string;
  fatherNameError: string;
  motherNameError: string;
  maleCountError: string;
  femaleCountError: string;
  maleMarriedCountError: string;
  femaleMarriedCountError: string;
  heightError: string;
  physicalStatusError: string;
  weightError: string;
  bodyTypeError: string;
  interCommunityError: string;
  locationError: string;
  cityError: string;
  stateError: string;
  addressError: string;
}

export const editeProfileValue = {
  interCommunity: '',
  height: '',
  physicalStatus: '',
  weight: '',
  bodyType: '',
  educationInstitution: '',
  Organization: '',
  smokingHabits: '',
  drinkingHabits: '',
  dietHabits: '',
  fatherOccupation: '',
  motherOccupation: '',
  education: '',
  employedIn: '',
  occupation: '',
  currency: 'INR',
  income: '',
  familyStatus: '',
  familyType: '',
  familyValue: '',
  ancestralOrigin: '',
  aboutMyself:
    'Kind-hearted professional with a passion for family and career. Enthusiastic about travel, food, and making meaningful connections. Seeking a life partner to share the journey of love and companionship.',
  fatherName: '',
  motherName: '',
  maleCount: '',
  femaleCount: '',
  maleMarriedCount: '',
  femaleMarriedCount: '',
  location: '',
  city: '',
  state: '',
  address: '',

  educationInstitutionError: '',
  OrganizationError: '',
  smokingHabitsError: '',
  drinkingHabitsError: '',
  dietHabitsError: '',
  fatherOccupationError: '',
  motherOccupationError: '',
  educationError: '',
  employedInError: '',
  occupationError: '',
  incomeError: '',
  familyStatusError: '',
  familyValueError: '',
  familyTypeError: '',
  ancestralOriginError: '',
  aboutMyselfError: '',
  fatherNameError: '',
  motherNameError: '',
  maleCountError: '',
  femaleCountError: '',
  maleMarriedCountError: '',
  femaleMarriedCountError: '',
  interCommunityError: '',
  heightError: '',
  physicalStatusError: '',
  weightError: '',
  bodyTypeError: '',
  locationError: '',
  cityError: '',
  stateError: '',
  addressError: '',
};